Move private serializers
